Multiple pie data labels: Typing, refactoring
